Fix problem in Latch
git-svn-id: svn+ssh://rubyforge.org/var/svn/cerberus@82 65aa75ef-ce15-0410-bc34-cdc86d5f77e6

+++ lib/cerberus/latch.rb
@@ -8,13 +8,10 @@ module Cerberus
         modif_time = File::Stat.new(lock_file).mtime
         ttl = options[:lock_ttl]
 
-        if ttl 
-          if modif_time < ttl.ago
-            File.delete(lock_file)
-          else
-            yield
-            return
-          end
+        if ttl and modif_time < ttl.ago
+          File.delete(lock_file)
+        else
+          return
         end
       end
